pyproject.toml
[tool.poetry]
name = "spoterfy"
version = "0.1.0"
description = "A Spotify CLI tool"
authors = ["Ali Macky"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.7"
requests = "^2.31.0"
urllib3 = "<2.0.0"
bottle = "^0.12.25"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
spoterfy = "spoterfy.cli:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 